fix: remove header menu panel
