--- apps/watch/src/components/Header/HeaderMenuPanel/index.ts
@@ -1,1 +1,0 @@
-export { HeaderMenuPanel } from './HeaderMenuPanel'
